// node/configure/type/tokyu-car-history.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type NoName1 = string;
export type NoName4 = string;

export interface NoName {
  view: View;
  validator: NoName2;
}
export interface View {
  init: NoName1;
}
export interface NoName2 {
  number: NoName3;
  register_start: NoName6;
  register_end: NoName8;
}
export interface NoName3 {
  regexp: NoName4;
  message: NoName5;
}
export interface NoName5 {
  [k: string]: unknown;
}
export interface NoName6 {
  message: NoName7;
}
export interface NoName7 {
  [k: string]: unknown;
}
export interface NoName8 {
  message: NoName9;
}
export interface NoName9 {
  [k: string]: unknown;
}